requirements.txt
lightning
timm
torch 
torchvision 
torchaudio
lightning
gradio
scikit-image
Pillow
numpy
joblib
scipy
pandas
matplotlib
wandb
torch_fidelity
opencv-python
pytorch_fid